Move Item Commands and Parsers to separate pkgs
