+++ src/main/java/seedu/foodrem/logic/commands/itemcommands/AddCommand.java
@@ -3,7 +3,7 @@ package seedu.foodrem.logic.commands.itemcommands;
 import static java.util.Objects.requireNonNull;
 import static seedu.foodrem.logic.parser.CliSyntax.PREFIX_ITEM_BOUGHT_DATE;
 import static seedu.foodrem.logic.parser.CliSyntax.PREFIX_ITEM_EXPIRY_DATE;
-import static seedu.foodrem.logic.parser.CliSyntax.PREFIX_ITEM_NAME;
+import static seedu.foodrem.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.foodrem.logic.parser.CliSyntax.PREFIX_ITEM_QUANTITY;
 import static seedu.foodrem.logic.parser.CliSyntax.PREFIX_ITEM_UNIT;
 
@@ -22,13 +22,13 @@ public class AddCommand extends Command {
 
     public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds an item to FoodRem. "
             + "Parameters: "
-            + PREFIX_ITEM_NAME + "NAME "
+            + PREFIX_NAME + "NAME "
             + PREFIX_ITEM_QUANTITY + "QUANTITY "
             + PREFIX_ITEM_UNIT + "UNIT "
             + PREFIX_ITEM_BOUGHT_DATE + "BOUGHT DATE "
             + PREFIX_ITEM_EXPIRY_DATE + "EXPIRY DATE "
             + "Example: " + COMMAND_WORD + " "
-            + PREFIX_ITEM_NAME + "Potatoes "
+            + PREFIX_NAME + "Potatoes "
             + PREFIX_ITEM_QUANTITY + "10 "
             + PREFIX_ITEM_UNIT + "kg "
             + PREFIX_ITEM_BOUGHT_DATE + "11-11-2022 "
