+++ src/main/java/seedu/foodrem/logic/parser/itemcommandparser/DeleteCommandParser.java
@@ -1,9 +1,11 @@
-package seedu.foodrem.logic.parser;
+package seedu.foodrem.logic.parser.itemcommandparser;
 
 import static seedu.foodrem.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.foodrem.commons.core.index.Index;
 import seedu.foodrem.logic.commands.itemcommands.DeleteCommand;
+import seedu.foodrem.logic.parser.Parser;
+import seedu.foodrem.logic.parser.ParserUtil;
 import seedu.foodrem.logic.parser.exceptions.ParseException;
 
 /**
