+++ src/main/java/seedu/foodrem/logic/parser/itemcommandparser/FindCommandParser.java
@@ -1,10 +1,11 @@
-package seedu.foodrem.logic.parser;
+package seedu.foodrem.logic.parser.itemcommandparser;
 
 import static seedu.foodrem.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import java.util.Arrays;
 
 import seedu.foodrem.logic.commands.itemcommands.FindCommand;
+import seedu.foodrem.logic.parser.Parser;
 import seedu.foodrem.logic.parser.exceptions.ParseException;
 import seedu.foodrem.model.item.NameContainsKeywordsPredicate;
